# Group1A4/Models/Instructor.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;

namespace Group1A4.Models
{
    public partial class Instructor
    {
        public int InstructorId { get; set; }
        public string InstructorFname { get; set; }
        public string InstructorLname { get; set; }
        public string InstructorEmail { get; set; }
        public int InstructorRating { get; set; }

        public virtual Class Class { get; set; }
    }
}