# Group1A4/Models/Student.cs
// <auto-generated> This file has been auto generated by EF Core Power Tools. </auto-generated>
#nullable disable
using System;
using System.Collections.Generic;

namespace Group1A4.Models
{
    public partial class Student
    {
        public int StudentId { get; set; }
        public string StudentFname { get; set; }
        public string StudentLname { get; set; }
        public string StudentEmail { get; set; }
        public string StudentZipcode { get; set; }

        public virtual Course Course { get; set; }
    }
}